+++ actors/multisig/events.go
@@ -9,7 +9,6 @@ import (
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/manifest"
 	filTypes "github.com/filecoin-project/lotus/chain/types"
-	"github.com/zondax/fil-parser/actors"
 	"github.com/zondax/fil-parser/parser"
 	"github.com/zondax/fil-parser/parser/helper"
 	"github.com/zondax/fil-parser/tools"
@@ -41,10 +40,6 @@ var cancelApproveTranslateMap = map[string]string{
 	parser.MethodApproveExported: parser.MethodApprove,
 	parser.MethodCancel:          parser.MethodCancel,
 	parser.MethodCancelExported:  parser.MethodCancel,
-}
-
-type EventGenerator interface {
-	GenerateMultisigEvents(ctx context.Context, transactions []*types.Transaction, tipsetCid string, tipsetKey filTypes.TipSetKey) (*types.MultisigEvents, error)
 }
 
 type eventGenerator struct {
@@ -192,7 +187,7 @@ func (eg *eventGenerator) processNestedParams(ctx context.Context, params map[st
 }
 
 func (eg *eventGenerator) createMultisigInfo(ctx context.Context, tx *types.Transaction, tipsetCid string) (*types.MultisigInfo, error) {
-	value, err := actors.ParseMultisigMetadata(tx.TxType, tx.TxMetadata)
+	value, err := ParseMultisigMetadata(tx.TxType, tx.TxMetadata)
 	if err != nil {
 		eg.logger.Sugar().Error(ctx, fmt.Sprintf("Multisig error parsing metadata: %s", err.Error()))
 		value = tx.TxMetadata // if there is an error then we need to store the raw metadata
