+++ actors/v11/multisig.go
@@ -1,60 +1,48 @@
 package v11
 
 import (
-	"bytes"
-
-	"github.com/filecoin-project/go-state-types/builtin/v8/multisig"
+	multisigV11 "github.com/filecoin-project/go-state-types/builtin/v11/multisig"
 	filTypes "github.com/filecoin-project/lotus/chain/types"
+	"github.com/zondax/fil-parser/actors/multisig"
 	"github.com/zondax/fil-parser/parser"
 )
 
-func ParseMultisig(txType string, msg *parser.LotusMessage, msgRct *parser.LotusMessageReceipt, height int64, key filTypes.TipSetKey) (map[string]interface{}, error) {
+type V11MultisigParser struct {
+	service multisig.MultisigMethods
+}
+
+func NewV11MultisigParser(service multisig.MultisigMethods) multisig.MultisigParser {
+	return &V11MultisigParser{service: service}
+}
+
+func (p *V11MultisigParser) ParseMultisig(txType string, msg *parser.LotusMessage, msgRct *parser.LotusMessageReceipt, height int64, key filTypes.TipSetKey) (map[string]interface{}, error) {
 	switch txType {
 	case parser.MethodConstructor:
-		return msigConstructor(msg.Params)
+		return p.service.ParseMsigConstructor(msg.Params, new(multisigV11.ConstructorParams))
 	case parser.MethodChangeNumApprovalsThreshold:
-		return changeNumApprovalsThreshold(msg.Params)
+		return p.service.ParseChangeNumApprovalsThreshold(msg.Params, new(multisigV11.ChangeNumApprovalsThresholdParams))
 	case parser.MethodLockBalance:
-		return lockBalance(msg.Params)
-		//// TODO: the rest of the methods
+		return p.service.ParseLockBalance(msg.Params, new(multisigV11.LockBalanceParams))
+	case parser.MethodRemoveSigner:
+		return p.service.ParseRemoveSigner(msg, height, key, new(multisigV11.RemoveSignerParams))
+	case parser.MethodApprove:
+		return p.service.ParseApprove(msg, msgRct.Return, height, key, new(multisigV11.TxnIDParams), new(multisigV11.ApproveReturn))
+	case parser.MethodCancel:
+		return p.service.ParseCancel(msg, height, key, new(multisigV11.TxnIDParams))
+	case parser.MethodPropose:
+		return p.service.ParsePropose(msg.Params, msgRct.Return, new(multisigV11.ProposeParams), new(multisigV11.ProposeReturn))
+
+		// TODO: the rest of the methods
+		//case parser.MethodAddSigner:
+		//	return p.service.ParseAddSigner(msg.Params, new(multisigV11.AddSignerParams))
+		//case parser.MethodSwapSigner:
+		//	return p.service.ParseSwapSigner(msg.Params, new(multisigV11.SwapSignerParams))
+		// case parser.MethodAddVerifier:
+		// 	return p.service.ParseAddVerifier(msg.Params, new(multisigV11.AddVerifierParams))
+		// case parser.MethodRemoveVerifier:
+		// 	return p.service.ParseRemoveVerifier(msg.Params, new(multisigV11.RemoveVerifierParams))
+		// case parser.MethodUniversalReceiverHook:
+		// 	return p.service.ParseUniversalReceiverHook(msg.Params, new(multisigV11.UniversalReceiverHookParams))
 	}
 	return map[string]interface{}{}, parser.ErrUnknownMethod
 }
-
-func msigConstructor(raw []byte) (map[string]interface{}, error) {
-	metadata := make(map[string]interface{})
-	reader := bytes.NewReader(raw)
-	var proposeParams multisig.ConstructorParams
-	err := proposeParams.UnmarshalCBOR(reader)
-	if err != nil {
-		return metadata, err
-	}
-	metadata[parser.ParamsKey] = proposeParams
-	return metadata, nil
-}
-
-func changeNumApprovalsThreshold(raw []byte) (map[string]interface{}, error) {
-	metadata := make(map[string]interface{})
-	var params multisig.ChangeNumApprovalsThresholdParams
-	reader := bytes.NewReader(raw)
-	err := params.UnmarshalCBOR(reader)
-	if err != nil {
-		return metadata, err
-	}
-	metadata[parser.ParamsKey] = params
-	return metadata, nil
-}
-
-func lockBalance(raw []byte) (map[string]interface{}, error) {
-	metadata := make(map[string]interface{})
-	var params multisig.LockBalanceParams
-	reader := bytes.NewReader(raw)
-	err := params.UnmarshalCBOR(reader)
-	if err != nil {
-		return metadata, err
-	}
-	metadata[parser.ParamsKey] = params
-	return metadata, nil
-}
-
-// the rest of code...
